requirements.txt
pydantic==2.8.2
djangorestframework==3.15.2
setuptools==72.1.0